requirements.txt
tqdm==4.65.0
torch==1.13.1
numpy==1.24.2
Pillow==9.4.0
torchvision==0.14.1
torchsummary==1.5.1
opencv-python==4.7.0.72
PyYAML==6.0